backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
aiofiles==24.1.0

# LLM & Orchestration
langchain==0.3.13
langchain-openai==0.2.14
langchain-core==0.3.28
langgraph==0.2.60

# Schema validation
pydantic==2.10.4
pydantic-settings==2.7.0

# Resume parsing
pdfplumber==0.11.4
python-docx==1.1.2
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==11.1.0

# Export
openpyxl==3.1.5

# Caching
diskcache==5.6.3

# Retry
tenacity==9.0.0

# Logging
structlog==24.4.0

# Utilities
PyYAML==6.0.2
httpx==0.28.1
python-dotenv==1.0.1
